BACKEND: Issue #86 Refactor old Api code * Moved mbta to new API structure * Moved weather to new API structure * Updated and moved old test to new API Structure * Updated UI for reset refactor * Removed defunct controller and test * Minor API test refactor to build off base class
