+++ tests/AppBundle/Controller/Api/CalendarControllerTest.php
@@ -4,44 +4,15 @@ namespace Tests\AppBundle\Controller\Api;
 
 use AppBundle\Entity\Calendar;
 use Statusboard\ControllerHelpers\CalendarHelper;
-use Statusboard\Utility\Environment;
-use Symfony\Bundle\FrameworkBundle\Client;
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Symfony\Component\HttpFoundation\Response;
 
-class CalendarControllerTest extends WebTestCase {
+class CalendarControllerTest extends ApiBase {
 
     const ENDPOINT_CALENDAR = '/api/calendar';
     const ENDPOINT_EVENT = self::ENDPOINT_CALENDAR . "/event";
     const ENDPOINT_UPCOMING = self::ENDPOINT_CALENDAR . "/upcoming";
 
     const CONTENT_TYPE_JSON = ['CONTENT_TYPE' => 'application/json'];
-
-    /**
-     * @var Client that is not logged in to the site
-     */
-    protected $loggedInClient;
-
-    /**
-     * @var Client that is logged into the site
-     */
-    protected $loggedOutClient;
-
-    /**
-     * Setup the loggedIn and LoggedOut clients
-     */
-    protected function setUp() {
-        parent::setUp();
-        Environment::getType(Environment::TYPE_TEST);
-        $this->loggedOutClient = static::createClient();
-        $this->loggedInClient = static::createClient(
-            [],
-            [
-                'PHP_AUTH_USER' => 'test',
-                'PHP_AUTH_PW'   => 'test12345',
-            ]
-        );
-    }
 
     /**
      * @test
